docurag-endee/requirements.txt
# Vector Database
endee>=0.1.0

# Embeddings
sentence-transformers>=2.7.0
torch>=2.2.0

# Document Processing
pymupdf>=1.24.0          # PDF parsing
python-docx>=1.1.0       # DOCX parsing
markdown>=3.6            # Markdown parsing

# API Framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9   # File upload support
pydantic>=2.7.0
pydantic-settings>=2.3.0

# LLM Clients
google-generativeai>=0.5.0
httpx>=0.27.0             # Ollama HTTP calls

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
typer>=0.12.0             # CLI

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0             # TestClient